move all pages from the pages folder to the app folder
