+++ apps/next/app/current-time/head.tsx
@@ -1,0 +1,11 @@
+import { DefaultTags } from '@/ui/DefaultTags'
+
+export default function Head() {
+	return (
+		<>
+			<DefaultTags />
+			<title>Current Time</title>
+			<meta name="description" content="Current Time" />
+		</>
+	)
+}
