+++ apps/next/app/current-time/page.tsx
@@ -1,0 +1,27 @@
+'use client'
+
+import useCurrentTime from 'lib/hooks/useCurrentTime'
+import Header from 'shared/Header'
+import Content from '@/components/Content'
+import Section from '@/components/Section'
+
+const CurrentTimePage = () => {
+	const { hours, minutes, meridiem } = useCurrentTime()
+
+	return (
+		<Content>
+			<Header />
+			<Section size="large">
+				<style jsx>{`
+					.meridiem {
+						margin-bottom: -18px;
+					}
+				`}</style>
+				{hours}:{minutes}
+				<span className="text-3/6 meridiem">{meridiem}</span>
+			</Section>
+		</Content>
+	)
+}
+
+export default CurrentTimePage
