+++ apps/next/app/hooks/head.tsx
@@ -1,0 +1,11 @@
+import { DefaultTags } from '@/ui/DefaultTags'
+
+export default function Head() {
+	return (
+		<>
+			<DefaultTags />
+			<title>Hooks</title>
+			<meta name="description" content="Progress bar" />
+		</>
+	)
+}
