+++ apps/next/app/ms/head.tsx
@@ -1,0 +1,17 @@
+import { DefaultTags } from '@/ui/DefaultTags'
+
+export default function Head() {
+	return (
+		<>
+			<DefaultTags />
+			<title>ms</title>
+			<meta name="description" content="ms" />
+			<link
+				rel="preload"
+				href="/fonts/MuseoModerno-Light.ttf"
+				as="font"
+				crossOrigin=""
+			/>
+		</>
+	)
+}
