+++ apps/next/app/password-generator/head.tsx
@@ -1,0 +1,11 @@
+import { DefaultTags } from '@/ui/DefaultTags'
+
+export default function Head() {
+	return (
+		<>
+			<DefaultTags />
+			<title>Password Generator</title>
+			<meta name="description" content="Password Generator" />
+		</>
+	)
+}
