+++ apps/next/app/progress-bar/head.tsx
@@ -1,0 +1,11 @@
+import { DefaultTags } from '@/ui/DefaultTags'
+
+export default function Head() {
+	return (
+		<>
+			<DefaultTags />
+			<title>Progress Bar</title>
+			<meta name="description" content="Progress bar" />
+		</>
+	)
+}
