+++ apps/next/app/random-string/head.tsx
@@ -1,0 +1,11 @@
+import { DefaultTags } from '@/ui/DefaultTags'
+
+export default function Head() {
+	return (
+		<>
+			<DefaultTags />
+			<title>Random String</title>
+			<meta name="description" content="Genrates a random string" />
+		</>
+	)
+}
