+++ apps/next/app/random-string/layout.tsx
@@ -1,0 +1,17 @@
+'use client'
+
+import type { PropsWithChildren } from 'react'
+import Header from 'shared/Header'
+import Content from '@/components/Content'
+
+export default function Layout({ children }: PropsWithChildren<unknown>) {
+	console.log(children)
+
+	return (
+		<Content>
+			<Header></Header>
+			{children}
+			{/* {cloneElement(children, { ref: randomStringRef })} */}
+		</Content>
+	)
+}
