+++ apps/next/app/rsc.server/head.tsx
@@ -1,0 +1,11 @@
+import { DefaultTags } from '@/ui/DefaultTags'
+
+export default function Head() {
+	return (
+		<>
+			<DefaultTags />
+			<title>React Server Components</title>
+			<meta name="description" content="React Server Components" />
+		</>
+	)
+}
