+++ apps/next/app/slots/head.tsx
@@ -1,0 +1,11 @@
+import { DefaultTags } from '@/ui/DefaultTags'
+
+export default function Head() {
+	return (
+		<>
+			<DefaultTags />
+			<title>Slots WIP</title>
+			<meta name="description" content="Slots WIP" />
+		</>
+	)
+}
