+++ apps/next/app/two-way-auth-enter/head.tsx
@@ -1,0 +1,11 @@
+import { DefaultTags } from '@/ui/DefaultTags'
+
+export default function Head() {
+	return (
+		<>
+			<DefaultTags />
+			<title>Enter Two-Factor Authentication</title>
+			<meta name="description" content="Enter two-factor authentication" />
+		</>
+	)
+}
