+++ apps/next/app/two-way-auth-generate/head.tsx
@@ -1,0 +1,11 @@
+import { DefaultTags } from '@/ui/DefaultTags'
+
+export default function Head() {
+	return (
+		<>
+			<DefaultTags />
+			<title>Generate Two-Factor Authentication</title>
+			<meta name="description" content="Generate two-factor authentication" />
+		</>
+	)
+}
