+++ apps/next/app/two-way-auth-generate/page.tsx
@@ -1,0 +1,17 @@
+'use client'
+
+import TwoWayAuthGenerate from '@/components/TwoWayAuthGenerate'
+import Header from 'shared/Header'
+import Content from '@/components/Content'
+import Section from '@/components/Section'
+
+const TwoWayAuthGeneratePage = () => (
+	<Content>
+		<Header />
+		<Section direction="column" size="medium">
+			<TwoWayAuthGenerate />
+		</Section>
+	</Content>
+)
+
+export default TwoWayAuthGeneratePage
