+++ apps/next/pages/_document.tsx
@@ -1,13 +1,13 @@
 import { Html, Head, Main, NextScript } from 'next/document'
 
-const name = 'Emil Johansson'
+// const name = 'Emil Johansson'
 export const siteTitle = 'emiljohansson.dev'
 
 export default function Document() {
 	return (
 		<Html lang="en">
 			<Head>
-				<link
+				{/* <link
 					rel="apple-touch-icon"
 					sizes="180x180"
 					href="/images/logo/apple-touch-icon.png"
@@ -34,7 +34,7 @@ export default function Document() {
 					)}.png?theme=light&md=0&fontSize=75px&images=https%3A%2F%2Fassets.zeit.co%2Fimage%2Fupload%2Ffront%2Fassets%2Fdesign%2Fnextjs-black-logo.svg`}
 				/>
 				<meta name="og:title" content={siteTitle} />
-				<meta name="twitter:card" content="summary_large_image" />
+				<meta name="twitter:card" content="summary_large_image" /> */}
 			</Head>
 			<body className="dark:bg-black-rich dark:text-white">
 				<Main />
