+++ apps/next/tsconfig.json
@@ -39,7 +39,8 @@
     "*.d.ts",
     "**/*.ts",
     "**/*.tsx",
-    ".next/types/**/*.ts"
+    ".next/types/**/*.ts",
+    "/Users/emiljohansson/Sites/emiljohansson.dev/apps/next/.next/types/**/*.ts"
   ],
   "exclude": [
     "node_modules"
